backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
websockets>=12.0
google-genai>=0.2.0
opencv-python-headless>=4.9.0.80
python-dotenv>=1.0.1
